+++ sdk/go/pulumi/resource.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2018, Pulumi Corporation.
+// Copyright 2016-2020, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,7 +14,9 @@
 
 package pulumi
 
-import "reflect"
+import (
+	"reflect"
+)
 
 type (
 	// ID is a unique identifier assigned by a resource provider to a resource.