+++ CHANGELOG.md
@@ -3,6 +3,9 @@ CHANGELOG
 
 ## HEAD (Unreleased)
 
+- Add support for generating Fish completions
+  [#4401](https://github.com/pulumi/pulumi/pull/4401)
+
 - Support map-typed inputs in RegisterResource for Go SDK
   [#4522](https://github.com/pulumi/pulumi/pull/4522)